LRIS-31868 LRIS-31869 port of LRIS-23541 LRIS-27391 update int(s) to long(s) to fix compile and update TicketEntry to AccountEnvironment

+++ modules/dxp/apps/osb/osb-customer/osb-portlet/docroot/WEB-INF/service/com/liferay/osb/model/AccountEnvironmentConstants.java
@@ -235,7 +235,7 @@ public class AccountEnvironmentConstants {
 
 	public static final long ENV_DB_ORACLE_12C_RELEASE_1 = 28030;
 
-	public static final int ENV_DB_ORACLE_12C_RELEASE_2 = 28040;
+	public static final long ENV_DB_ORACLE_12C_RELEASE_2 = 28040;
 
 	public static final long ENV_DB_OTHER = 28008;
 
@@ -379,7 +379,7 @@ public class AccountEnvironmentConstants {
 
 	public static final long ENV_OS_WINDOWS_SERVER_2012 = 30030;
 
-	public static final int ENV_OS_WINDOWS_SERVER_2016 = 30043;
+	public static final long ENV_OS_WINDOWS_SERVER_2016 = 30043;
 
 	public static final long ENV_OS_WINDOWS_XP = 30003;
